ALS SNVs for B Parameter Training

Pre-processed GWAS summary statistics for Amyotrophic Lateral Sclerosis (ALS) from literature studies. This dataset has been cleaned and standardized for use in B parameter estimation (effect size as a function of MAF).

ALS_snvs_B_training

Format

A data frame with 75 rows (variants) and 9 columns:

rsID

SNP identifier (character)

CHR

Chromosome, autosomes only (numeric 1-22)

POS

Genomic position in GRCh37 (numeric)

MAF

Minor allele frequency, 0 < MAF \(\le\) 0.5 (numeric)

P

P-value, 0 < P \(\le\) 1 (numeric)

N

Sample size, N \(\ge\) 1000 (numeric)

BETA

Effect size as log odds ratio (numeric)

TRAIT_TYPE

"binary" for all variants (character)

STUDY

Study accession identifier (character)

Source

GWAS Catalog https://www.ebi.ac.uk/gwas/

Details

Processing

Prepared from raw ALS literature data using prepare_B_training_data with the following quality control:

Autosomal variants only (CHR 1-22)

Minimum sample size: N \(\ge\) 1000

Non-diagnostic traits excluded (age of onset, survival, C9orf72 interaction)

Nicolas A et al. (2018) study excluded (see note below)

Duplicates removed (kept variant with largest N)

Rows with NA in MAF, P, N, or BETA removed

The Nicolas A study is excluded because its variants overlap substantially with other included studies but report from a different analysis model. This validated 75-SNV dataset produces B estimates with R-squared = 0.50 using the beta method, and correlation > 0.9999 with the original reference formula.

Provenance: prepared from the curated 75-SNV ALS set; the raw input inst/extdata/ALS-known-SNPs-raw.xlsx ships with the package.

See also

Examples

# Load the data
data(ALS_snvs_B_training)

# Examine structure
str(ALS_snvs_B_training)
#> 'data.frame':	75 obs. of  9 variables:
#>  $ rsID      : chr  "rs6703183" "rs16856202" "rs873917" "rs12145183" ...
#>  $ CHR       : num  1 1 1 1 10 10 10 11 11 11 ...
#>  $ POS       : num  2.10e+08 2.32e+08 4.01e+07 7.53e+07 1.14e+08 ...
#>  $ MAF       : num  0.3416 0.04 0.29 0.0762 0.348 ...
#>  $ P         : num  1e-08 8e-06 8e-06 2e-14 7e-10 ...
#>  $ N         : num  4654 4079 4079 4654 138086 ...
#>  $ BETA      : num  0.262 0.693 0.148 0.438 0.06 ...
#>  $ TRAIT_TYPE: chr  "binary" "binary" "binary" "binary" ...
#>  $ STUDY     : chr  "GCST009529" "GCST000406" "GCST000406" "GCST009529" ...
#>  - attr(*, "column_mappings")=List of 11
#>   ..$ BETA        : chr "Beta_numeric"
#>   ..$ MAF         : chr "RAF"
#>   ..$ P           : chr "P-value"
#>   ..$ N           : chr "Sample Size"
#>   ..$ rsID        : chr "rsID"
#>   ..$ CHR         : chr "chr"
#>   ..$ POS         : chr "pos_GRCh37"
#>   ..$ STUDY       : chr "Study accession"
#>   ..$ PMID        : chr "PubMed ID"
#>   ..$ FIRST_AUTHOR: chr "First Author"
#>   ..$ OR          : chr "OR"
#>  - attr(*, "mapping_sources")=List of 11
#>   ..$ BETA        : chr "user-specified"
#>   ..$ MAF         : chr "user-specified"
#>   ..$ P           : chr "user-specified"
#>   ..$ N           : chr "user-specified"
#>   ..$ rsID        : chr "user-specified"
#>   ..$ CHR         : chr "user-specified"
#>   ..$ POS         : chr "user-specified"
#>   ..$ STUDY       : chr "user-specified"
#>   ..$ PMID        : chr "user-specified"
#>   ..$ FIRST_AUTHOR: chr "user-specified"
#>   ..$ OR          : chr "auto-detected"
head(ALS_snvs_B_training)
#>         rsID CHR       POS    MAF     P      N       BETA TRAIT_TYPE
#> 1  rs6703183   1 209712889 0.3416 1e-08   4654 0.26236426     binary
#> 2 rs16856202   1 232155151 0.0400 8e-06   4079 0.69314718     binary
#> 3   rs873917   1  40132795 0.2900 8e-06   4079 0.14842001     binary
#> 4 rs12145183   1  75272480 0.0762 2e-14   4654 0.43825493     binary
#> 5 rs58854276  10 114145044 0.3480 7e-10 138086 0.06000000     binary
#> 6  rs3736947  10 114169190 0.3500 8e-11  84343 0.08342161     binary
#>          STUDY
#> 1   GCST009529
#> 2   GCST000406
#> 3   GCST000406
#> 4   GCST009529
#> 5 GCST90027163
#> 6   GCST010857

if (FALSE) { # \dontrun{
# Use in B estimation
B <- get_B(
  training_trait = "binary",
  training_MAF = ALS_snvs_B_training$MAF,
  training_BETA = ALS_snvs_B_training$BETA,
  target_trait = "binary",
  target_MAF = seq(0.01, 0.4, by = 0.05),
  target_case_prop = 0.5
)
} # }