BMD SNVs for B Parameter Training

Pre-processed GWAS summary statistics for Bone Mineral Density (BMD) from large-scale literature studies. This dataset has been cleaned and standardized for use in B parameter estimation (effect size as a function of MAF).

BMD_snvs_B_training

Format

A data frame with 1790 rows (variants) and 10 columns:

rsID

SNP identifier (character)

CHR

Chromosome (numeric)

POS

Genomic position (numeric)

MAF

Minor allele frequency, 0 < MAF \(\le\) 0.5 (numeric)

P

P-value, 0 < P \(\le\) 1 (numeric)

P_mlog10

Negative log10 p-value, i.e., -log10(P) (numeric)

N

Sample size, N \(\ge\) 1000 (numeric)

BETA

Effect size (numeric, may contain NA)

TRAIT_TYPE

"continuous" for all variants (character)

STUDY

Study identifier (character)

Source

GWAS Catalog https://www.ebi.ac.uk/gwas/

Details

Processing

Prepared from raw BMD/osteoporosis literature data (CSV from GWAS Catalog) using prepare_B_training_data with the following quality control:

Three high-quality studies included: "An atlas of genetic influences on osteoporosis" (Morris 2019), "Identification of 153 new loci", "Identification of 613 new loci"

Binary trait "bone fracture" excluded

Minimum sample size: N \(\ge\) 1000

Rows with NA in MAF, P, or N removed

Duplicates NOT removed (retains variants from different studies)

This is a continuous trait dataset. For cross-trait prediction (e.g., BMD to osteoporosis risk), use the p-value method in get_B.

Provenance: prepared from the curated BMD/osteoporosis literature summary statistics (raw inputs are maintained with the source study, not bundled).

References

Morris JA, Kemp JP, Youlten SE, et al. (2019). An atlas of genetic influences on osteoporosis in humans and mice. Nature Genetics, 51(2), 258-266. doi:10.1038/s41588-018-0302-x

See also

Examples

# Load the data
data(BMD_snvs_B_training)

# Examine structure
str(BMD_snvs_B_training)
#> 'data.frame':	1790 obs. of  10 variables:
#>  $ rsID      : chr  "1:114494567_CT_C" "1:220189136_TA_T" "1:85296916_TG_T" "1:8657844_AT_A" ...
#>  $ CHR       : int  1 1 1 1 1 10 10 10 10 10 ...
#>  $ POS       : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ MAF       : num  0.374 0.178 0.115 0.166 0.227 ...
#>  $ P         : num  2e-11 3e-36 5e-09 1e-23 2e-12 ...
#>  $ P_mlog10  : num  10.7 35.5 8.3 23 11.7 ...
#>  $ N         : num  394929 394929 394929 394929 394929 ...
#>  $ BETA      : num  0.0148 0.0338 -0.0183 0.0276 -0.0168 ...
#>  $ TRAIT_TYPE: chr  "continuous" "continuous" "continuous" "continuous" ...
#>  $ STUDY     : chr  "Identification of 613 new loci associated with heel bone mineral density and a polygenic risk score for bone mi"| __truncated__ "Identification of 613 new loci associated with heel bone mineral density and a polygenic risk score for bone mi"| __truncated__ "Identification of 613 new loci associated with heel bone mineral density and a polygenic risk score for bone mi"| __truncated__ "Identification of 613 new loci associated with heel bone mineral density and a polygenic risk score for bone mi"| __truncated__ ...
head(BMD_snvs_B_training)
#>                        rsID CHR POS      MAF     P P_mlog10      N       BETA
#> 1          1:114494567_CT_C   1  NA 0.373892 2e-11 10.69897 394929  0.0147978
#> 2          1:220189136_TA_T   1  NA 0.178305 3e-36 35.52288 394929  0.0337965
#> 4           1:85296916_TG_T   1  NA 0.114804 5e-09  8.30103 394929 -0.0182615
#> 5            1:8657844_AT_A   1  NA 0.165803 1e-23 23.00000 394929  0.0275657
#> 6      1:98403226_TTTGTTG_T   1  NA 0.226728 2e-12 11.69897 394929 -0.0167524
#> 7 10:123694303_AGTTATAAAC_A  10  NA 0.086143 2e-18 17.69897 394929  0.0316996
#>   TRAIT_TYPE
#> 1 continuous
#> 2 continuous
#> 4 continuous
#> 5 continuous
#> 6 continuous
#> 7 continuous
#>                                                                                                                                                      STUDY
#> 1 Identification of 613 new loci associated with heel bone mineral density and a polygenic risk score for bone mineral density, osteoporosis and fracture.
#> 2 Identification of 613 new loci associated with heel bone mineral density and a polygenic risk score for bone mineral density, osteoporosis and fracture.
#> 4 Identification of 613 new loci associated with heel bone mineral density and a polygenic risk score for bone mineral density, osteoporosis and fracture.
#> 5 Identification of 613 new loci associated with heel bone mineral density and a polygenic risk score for bone mineral density, osteoporosis and fracture.
#> 6 Identification of 613 new loci associated with heel bone mineral density and a polygenic risk score for bone mineral density, osteoporosis and fracture.
#> 7 Identification of 613 new loci associated with heel bone mineral density and a polygenic risk score for bone mineral density, osteoporosis and fracture.

if (FALSE) { # \dontrun{
# Use in B estimation (continuous trait, p-value method)
B <- get_B(
  training_trait = "continuous",
  training_MAF = BMD_snvs_B_training$MAF,
  training_P = BMD_snvs_B_training$P,
  training_N = BMD_snvs_B_training$N,
  target_trait = "binary",      # Cross-trait: BMD -> osteoporosis risk
  target_MAF = seq(0.01, 0.4, by = 0.05),
  target_case_prop = 0.20
)
} # }